Move to ESM, first change for inter moldule events
Moved everything over to ESM syntax.
Added a test case for the new consolidated event emitter.
Added the new ConsolidatedListener Class for listening to multiple event emitters for events (will be required for the inter module events).
First step to setting up the inter module events.
Renamed the etc dir to "core" to highlight that they are all files required for the core functionality of the bot