Misc Fixes
Fixed an incorrect usage of lodash on the welcome module.
Escaped user created regex in the filter module (to prevent malicious regex from being created), the common filters can be used instead to provide flexible regex.
Modified the permissions check to be more robust for users that aren't members? not sure when this would happen, but have noticed that it is possible based on bot logs.