Changed config file parsing + Misc Changes
Added a centralised config file parsing module. Modified the emoji controller to check for reactions in the cache before deleting the reaction (this should hopefully stop it from attempting to remove reactions from a deleted message). Added a token check on startup (to notify the user that there is no token)